Update Spawn Handler
There's a mix of a few things, the first being the inclusion of a checkReplacements method that, like the rest of the spawn handlers, will now run the appropriate spawn replacement logic checks if the entry is marked to replace (the default).  This also incurs a change to what's allowed to be used as a spawn slot, namely the "random" or "spawner" classes that specifically place the weapon/magazine/etc., rather than the weapon or magazine itself.  This has the added benefit of not spawning/replacing magazines meant for different weapons in the same spot.

The other main thing changed is the addition of a few missing entries to the backpack blacklist, as well as fixing up some menu localization entries.